requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
certifi==2026.2.25
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.124.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
joblib==1.4.2
numpy==1.24.4
packaging==26.0
pandas==2.0.3
pluggy==1.5.0
psutil==7.2.2
pydantic==2.10.6
pydantic-core==2.27.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2026.1.post1
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
threadpoolctl==3.5.0
tomli==2.4.1
typing-extensions==4.13.2
tzdata==2025.3
uvicorn==0.33.0
